Split up http_api into files
