+++ crates/librqbit/src/lib.rs
@@ -49,8 +49,10 @@ pub mod file_info;
 mod file_ops;
 #[cfg(feature = "http-api")]
 pub mod http_api;
-#[cfg(feature = "http-api")]
+#[cfg(feature = "http-api-client")]
 pub mod http_api_client;
+#[cfg(any(feature = "http-api", feature = "http-api-client"))]
+pub mod http_api_types;
 pub mod limits;
 mod merge_streams;
 mod peer_connection;
